+++ test/gurka/test_gtfs.cc
@@ -1,7 +1,10 @@
 #include "gurka.h"
 #include "just_gtfs/just_gtfs.h"
 
-#include "string.h"
+#include "mjolnir/convert_transit.h"
+#include "mjolnir/ingest_transit.h"
+#include "proto/transit.pb.h"
+#include "test.h"
 #include <gtest/gtest.h>
 
 using namespace gtfs;
@@ -19,12 +22,18 @@ boost::property_tree::ptree get_config() {
 TEST(GtfsExample, WriteGtfs) {
   auto pt = get_config();
 
-  const std::string tripOneID = "bar";
-  const std::string tripTwoID = "barbar";
-  const std::string stopOneID = "foo";
-  const std::string stopTwoID = "foo2";
-  const std::string shapeOneID = "square";
-  const std::string serviceOneID = "bon";
+  const std::string tripOneID = "10";
+  const std::string tripTwoID = "11";
+  const std::string stopOneID = "7";
+  const std::string stopTwoID = "8";
+  const std::string stopThreeID = "19";
+  const std::string shapeOneID = "5";
+  const std::string serviceOneID = "9";
+  const int serviceStartDate = 20220131;
+  const int serviceEndDate = 20230131;
+  const int addedDate = 20220202;
+  const int removedDate = 20220203;
+  const int headwaySec = 1800;
   Feed feed;
 
   std::string path_directory = pt.get<std::string>("mjolnir.transit_feeds_dir") +
@@ -59,9 +68,18 @@ TEST(GtfsExample, WriteGtfs) {
 
   feed.write_stops(path_directory);
 
+  struct gtfs::Stop thirdStop {
+    .stop_id = stopThreeID, .stop_name = gtfs::Text("THIRD STOP"), .coordinates_present = true,
+    .stop_lat = 5, .stop_lon = 5, .parent_station = "1",
+    .location_type = gtfs::StopLocationType::StopOrPlatform, .stop_timezone = "America/Toronto",
+    .wheelchair_boarding = "1",
+  };
+  feed.add_stop(thirdStop);
+  feed.write_stops(("test/data/gtfs_feeds/toronto"));
+
   // write routes.txt
   struct Route lineOne {
-    .route_id = "baz", .route_type = RouteType::Subway, .route_short_name = "ba",
+    .route_id = "2", .route_type = RouteType::Subway, .route_short_name = "ba",
     .route_long_name = "bababa", .route_desc = "this is the first route", .route_color = "blue",
     .route_text_color = "black",
   };
@@ -71,8 +89,8 @@ TEST(GtfsExample, WriteGtfs) {
 
   // write trips.txt
   struct gtfs::Trip tripOne {
-    .route_id = "baz", .service_id = serviceOneID, .trip_id = tripOneID, .trip_headsign = "haha",
-    .block_id = "block", .shape_id = shapeOneID, .wheelchair_accessible = gtfs::TripAccess::Yes,
+    .route_id = "2", .service_id = serviceOneID, .trip_id = tripOneID, .trip_headsign = "haha",
+    .block_id = "3", .shape_id = shapeOneID, .wheelchair_accessible = gtfs::TripAccess::Yes,
     .bikes_allowed = gtfs::TripAccess::No,
   };
   feed.add_trip(tripOne);
@@ -86,14 +104,15 @@ TEST(GtfsExample, WriteGtfs) {
   feed.write_trips(path_directory);
 
   // write stop_times.txt
-  for (int i = 0; i < 4; i++) {
+  for (int i = 0; i < 6; i++) {
+    std::string stopIds[3] = {stopOneID, stopTwoID, stopThreeID};
     struct StopTime stopTime {
       .trip_id = "", .stop_id = "", .stop_sequence = 0, .arrival_time = Time("6:00:00"),
       .departure_time = Time("6:00:00"), .stop_headsign = "head", .shape_dist_traveled = 0.0,
       .timepoint = gtfs::StopTimePoint::Exact,
     };
-    stopTime.stop_id = (i % 2 == 0) ? stopOneID : stopTwoID;
-    stopTime.trip_id = (i < 2) ? tripOneID : tripTwoID;
+    stopTime.stop_id = stopIds[i % 3];
+    stopTime.trip_id = (i < 3) ? tripOneID : tripTwoID;
     feed.add_stop_time(stopTime);
   }
 
@@ -129,7 +148,8 @@ TEST(GtfsExample, WriteGtfs) {
   // write shapes.txt
   for (size_t i = 0; i < 4; i++) {
     struct ShapePoint shapePointTest {
-      .shape_id = shapeOneID, .shape_pt_lat = 0.2, .shape_pt_lon = 0.2, .shape_pt_sequence = (i + 1),
+      .shape_id = shapeOneID, .shape_pt_lat = i + 0.2, .shape_pt_lon = i + 0.2,
+      .shape_pt_sequence = (i + 1),
     };
     feed.add_shape(shapePointTest);
   }
@@ -139,11 +159,11 @@ TEST(GtfsExample, WriteGtfs) {
   // write frequencies.txt
   struct Frequency freqBased {
     .trip_id = tripOneID, .start_time = Time(0, 0, 0), .end_time = Time(2, 0, 0),
-    .headway_secs = 1800, .exact_times = gtfs::FrequencyTripService::FrequencyBased,
+    .headway_secs = headwaySec, .exact_times = gtfs::FrequencyTripService::FrequencyBased,
   };
   struct Frequency schedBased {
     .trip_id = tripTwoID, .start_time = Time(0, 0, 0), .end_time = Time(2, 0, 0),
-    .headway_secs = 1800, .exact_times = gtfs::FrequencyTripService::ScheduleBased,
+    .headway_secs = headwaySec, .exact_times = gtfs::FrequencyTripService::ScheduleBased,
   };
 
   feed.add_frequency(freqBased);
@@ -157,11 +177,11 @@ TEST(GtfsExample, WriteGtfs) {
   // make sure files are actually written
 
   const auto& trips = feed_reader.get_trips();
-  EXPECT_EQ(trips.size(), 1);
+  EXPECT_EQ(trips.size(), 2);
   EXPECT_EQ(trips[0].trip_id, tripOneID);
 
   const auto& stops = feed_reader.get_stops();
-  EXPECT_EQ(stops.size(), 2);
+  EXPECT_EQ(stops.size(), 3);
   EXPECT_EQ(stops[1].stop_id, stopTwoID);
 
   const auto& shapes = feed_reader.get_shapes();
@@ -186,15 +206,16 @@ TEST(GtfsExample, WriteGtfs) {
 }
 
 TEST(GtfsExample, MakeProto) {
-  auto pt = path_directory;
-  filesystem::create_directories(VALHALLA_BUILD_DIR "test/data/transit_test");
-  filesystem::create_directories(VALHALLA_BUILD_DIR "test/data/transit_tiles");
+  filesystem::create_directories("test/data/transit_test");
+  filesystem::create_directories("test/data/transit_tiles");
+  auto pt = get_config();
 
   // constants written in the last function
   const std::string tripOneID = "10";
   const std::string tripTwoID = "11";
   const std::string stopOneID = "7";
   const std::string stopTwoID = "8";
+  const std::string stopThreeID = "19";
   const std::string shapeOneID = "5";
   const std::string serviceOneID = "9";
   const int serviceStartDate = 20220131;
@@ -211,10 +232,11 @@ TEST(GtfsExample, MakeProto) {
   valhalla::mjolnir::stitch_transit(pt, dangling_tiles);
   // call the two functions, in main valhalla_ingest-transit
   // it's gonna write protobufs
-  filesystem::recursive_directory_iterator transit_file_itr(VALHALLA_BUILD_DIR
-                                                            "test/data/transit_tiles");
+  filesystem::recursive_directory_iterator transit_file_itr("test/data/transit_tiles");
   filesystem::recursive_directory_iterator end_file_itr;
 
+  std::unordered_set<std::string> stops;
+  std::unordered_set<std::string> stop_pairs;
   // for each pbf.
   for (; transit_file_itr != end_file_itr; ++transit_file_itr) {
     if (filesystem::is_regular_file(transit_file_itr->path())) {
@@ -229,15 +251,19 @@ TEST(GtfsExample, MakeProto) {
       EXPECT_EQ(transit.shapes_size(), 1);
       EXPECT_EQ(transit.shapes(0).shape_id(), stoi(shapeOneID));
       // stop(node) info
-      EXPECT_EQ(transit.nodes_size(), 2);
-      EXPECT_EQ(transit.nodes(0).onestop_id(), stopOneID);
+      for (int i = 0; i < transit.nodes_size(); i++) {
+        stops.insert(transit.nodes(i).onestop_id());
+      }
+
       // routes info
       EXPECT_EQ(transit.routes_size(), 1);
       EXPECT_EQ(transit.routes(0).onestop_id(), "2");
 
       // stop_pair info
-      EXPECT_EQ(transit.stop_pairs_size(), 2);
-      EXPECT_EQ(transit.stop_pairs(0).origin_onestop_id(), stopOneID);
+      for (int i = 0; i < transit.stop_pairs_size(); i++) {
+        stop_pairs.insert(transit.stop_pairs(i).origin_onestop_id());
+        stop_pairs.insert(transit.stop_pairs(i).destination_onestop_id());
+      }
 
       // calendar information
       EXPECT_EQ(transit.stop_pairs(0).service_start_date(), serviceStartDate);
@@ -254,13 +280,20 @@ TEST(GtfsExample, MakeProto) {
       EXPECT_EQ(transit.stop_pairs(1).frequency_headway_seconds(), headwaySec);
     }
   }
+  EXPECT_EQ(stops.size(), 3);
+  std::string stopIds[3] = {stopOneID, stopTwoID, stopThreeID};
+
+  for (const auto& stopID : stopIds) {
+    EXPECT_EQ(*stops.find((stopID)), stopID);
+    EXPECT_EQ(*stop_pairs.find((stopID)), stopID);
+  }
 }
 
 TEST(GtfsExample, MakeTile) {
   auto pt = get_config;
   filesystem::create_directories(VALHALLA_BUILD_DIR "test/data/level3_tile_dir");
 
-  auto all_tiles = valhalla::mjolnir::convert_transit(pt);
+  auto all_tiles = valhalla::mjolnir::convert_transit(*pt);
 
   //  // files are already going to be written from
   //  filesystem::recursive_directory_iterator transit_file_itr("test/data/transit_tiles");
@@ -273,3 +306,5 @@ TEST(GtfsExample, MakeTile) {
   //    }
   //  }
 }
+
+// make sure all the data is inside -> which we can test using the same tests as above